Get validation samples uniformly from all shards
The validation set used to be generated from the combined array of
all samples, irrespective of which shard they belong to.
This could result in an unequal distribution of validation samples
between workers.
Now, the specified fraction of samples is picked from every data file
separately, which ensures a uniform distribution as long as the shards
themselves have the same size.